requirements.txt
asgiref==3.8.1
Django==4.2
django-formtools==2.5.1
django-nested-admin==4.1.1
python-monkey-business==1.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
Unidecode==1.3.8
